API adds role to the user
